# mnemonic/internal/http/ui/openapi.yaml
openapi: 3.0.3
info:
  title: Skillgrid Mnemonic HTTP API
  description: |-
    Persistent memory, code index, and web research cache for AI agents.

    Write routes (POST) optionally require an `Authorization: Bearer <token>`
    header when the server is started with `SKILLGRID_HTTP_TOKEN` set.
    Most routes take a `project` query parameter naming the project store
    (see `GET /projects` for available IDs).
  version: "0.1.0"
servers:
  - url: /
...
paths:
...
  /sessions/{id}/events:
    get:
      tags: [session]
      operationId: sessionToolEvents
      summary: One session's tool-call timeline
      description: |-
        Harness tool calls for the session, newest first. Takes the same
        filters as GET /events. Sensitive paths keep their metadata but never
        return a preview. Policy decision rows (result blocked, warned, or
        guided) carry "rule: message" as the preview.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: project, in: query, required: true, schema: { type: string } }
        - { $ref: "#/components/parameters/ToolAgent" }
        - { $ref: "#/components/parameters/ToolAction" }
        - { $ref: "#/components/parameters/ToolName" }
        - { $ref: "#/components/parameters/ToolFile" }
        - { $ref: "#/components/parameters/ToolCommand" }
        - { $ref: "#/components/parameters/ToolSince" }
        - { name: limit, in: query, schema: { type: integer, default: 200, maximum: 2000 } }
      responses:
        "200":
          description: Tool events
          content:
            application/json:
              schema:
                type: object
                properties:
                  project: { type: string }
                  sessionId: { type: string }
                  events:
                    type: array
                    items: { $ref: "#/components/schemas/ToolEvent" }
                  observations:
                    type: array
                    description: Live observations for the session (pinned first, newest first)
                    items: { $ref: "#/components/schemas/SessionObservationFeedItem" }
        "400": { $ref: "#/components/responses/Bad" }
...
  /events:
    get:
      tags: [session]
      operationId: toolEvents
      summary: Query tool calls across every session
      description: |-
        The Gryph-style query behind the Sessions filter bar and
        `skillgrid logs`. File and command filters are globs: * and ** match
        any run, ? matches one character.
      parameters:
        - { name: project, in: query, required: true, schema: { type: string } }
        - { name: session, in: query, schema: { type: string } }
        - { $ref: "#/components/parameters/ToolAgent" }
        - { $ref: "#/components/parameters/ToolAction" }
        - { $ref: "#/components/parameters/ToolName" }
        - { $ref: "#/components/parameters/ToolFile" }
        - { $ref: "#/components/parameters/ToolCommand" }
        - { $ref: "#/components/parameters/ToolSince" }
        - { name: until, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 200, maximum: 2000 } }
      responses:
        "200":
          description: Tool events, newest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items: { $ref: "#/components/schemas/ToolEvent" }
        "400": { $ref: "#/components/responses/Bad" }
...
  /context:
    get:
      tags: [memory]
      operationId: context
      summary: Recent session context
      parameters:
        - name: project
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Recent sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sessions"
        "400":
          $ref: "#/components/responses/Bad"
        "500":
          $ref: "#/components/responses/Internal"
...
  /web/search:
    get:
      tags: [web]
      operationId: webSearch
      summary: Full-text search over cached web snapshots
      parameters:
        - name: project
          in: query
          required: true
          schema:
            type: string
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: source
          in: query
          required: false
          schema:
            type: string
        - name: fresh_only
          in: query
          required: false
          schema:
            type: string
            enum: ["true", "false"]
            default: "true"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Matching entries
          content:
            application/json:
              schema:
                type: object
                required: [entries]
                properties:
                  entries:
                    type: array
                    items:
                      $ref: "#/components/schemas/WebHit"
        "400":
          $ref: "#/components/responses/Bad"
        "500":
          $ref: "#/components/responses/Internal"
...
  /activity/events:
    get:
      tags: [activity]
      operationId: activityEvents
      summary: Agent activity feed (newest first)
      description: |-
        Returns the project's agent activity (observations) newest first.
        There is no events table — activity is modeled as observations.
        limit is clamped to 1..500. Unknown project → 200 with an empty feed.
      parameters:
        - name: project
          in: query
          required: true
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
      responses:
        "200":
          description: Activity events
          content:
            application/json:
              schema:
                type: object
                properties:
                  project: { type: string }
                  events:
                    type: array
                    items: { $ref: "#/components/schemas/ActivityEvent" }
                  limit: { type: integer }
        "400": { $ref: "#/components/responses/Bad" }
...
components:
  responses:
    Bad:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
    Internal:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
  parameters:
    ToolAgent: { name: agent, in: query, schema: { type: string, example: cursor } }
    ToolAction:
      name: action
      in: query
      schema: { type: string, enum: [file_read, file_write, command_exec, tool_use] }
    ToolName: { name: tool, in: query, schema: { type: string } }
    ToolFile: { name: file, in: query, description: "path glob", schema: { type: string } }
    ToolCommand: { name: command, in: query, description: "command glob", schema: { type: string } }
    ToolSince: { name: since, in: query, description: "RFC3339 or 30m, 24h, 7d, 1w", schema: { type: string } }
  schemas:
    ToolEvent:
      type: object
      properties:
        id: { type: integer, format: int64 }
        ts: { type: string }
        sessionId: { type: string }
        sequence: { type: integer }
        agent: { type: string }
        action: { type: string }
        tool: { type: string }
        path: { type: string }
        command: { type: string }
        result: { type: string, description: "success, error, blocked, warned, or guided" }
        sensitive: { type: boolean }
        preview: { type: string }
        mcp: { type: boolean }
    HTTPError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    SessionObservationFeedItem:
      type: object
      properties:
        id: { type: integer, format: int64 }
        type: { type: string }
        title: { type: string }
        created_at: { type: string, format: date-time }
        tokens: { type: integer, description: Heuristic token estimate (title + content) }
        pinned: { type: boolean }
...
    Sessions:
      type: object
      required: [sessions]
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/Session"
...
    Session:
      type: object
      properties:
        id:
          type: string
        project:
          type: string
        directory:
          type: string
        title:
          type: string
          nullable: true
          description: Human-readable session name (optional); shown in the dashboard session list
        started_at:
          type: string
        ended_at:
          type: string
        summary:
          type: string
        status:
          type: string
...
    WebHit:
      type: object
      properties:
        id:
          type: integer
          format: int64
        source:
          type: string
        title:
          type: string
        query:
          type: string
        url:
          type: string
        library_id:
          type: string
        fetched_at:
          type: string
        expires_at:
          type: string